Change project API so that the user must own the project to perform actions on it or read it.

+++ Controllers/ProjectController.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using ProjectManagementApp.Contracts;
 using ProjectManagementApp.Models;
-using ProjectManagementApp.Services;
+using System.Security.Claims;
 
 namespace ProjectManagementApp.Controllers
 {
@@ -12,17 +12,19 @@ namespace ProjectManagementApp.Controllers
     public class ProjectController : ControllerBase
     {
         private readonly IProjectService _projectService;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public ProjectController(IProjectService projectService)
+        public ProjectController(IProjectService projectService, IHttpContextAccessor httpContextAccessor)
         {
             _projectService = projectService;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         // GET: api/projects
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Project>>> GetProjects()
         {
-            var projects = await _projectService.GetProjectsAsync();
+            var projects = await _projectService.GetProjectsWithUserIdAsync(GetUserId());
 
             return Ok(projects);
         }
@@ -31,11 +33,16 @@ namespace ProjectManagementApp.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<Project>> GetProject(int id)
         {
-            var project = await _projectService.GetProjectByIdAsync(id);
+            var project = await _projectService.GetProjectWithIdAsync(id);
 
             if (project == null)
             {
                 return NotFound();
+            }
+
+            if (project.UserId != GetUserId())
+            {
+                return Forbid();
             }
 
             return Ok(project);
@@ -45,6 +52,8 @@ namespace ProjectManagementApp.Controllers
         [HttpPost]
         public async Task<ActionResult<Project>> PostProject(Project project)
         {
+            project.UserId = GetUserId();
+
             var projectId = await _projectService.CreateProjectAsync(project);
 
             return CreatedAtAction("GetProject", new { id = projectId }, project);
@@ -59,6 +68,13 @@ namespace ProjectManagementApp.Controllers
                 return BadRequest();
             }
 
+            // FIXME!!! Make sure to check the project that is stored in the database to make sure that it has a 
+            // matching UserId. Also, make sure the user can't send a new UserId.
+            if (project.UserId != GetUserId())
+            {
+                return Forbid();
+            }
+
             await _projectService.UpdateProjectAsync(project);
 
             return NoContent();
@@ -68,6 +84,18 @@ namespace ProjectManagementApp.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteProject(int id)
         {
+            var project = await _projectService.GetProjectWithIdAsync(id);
+
+            if (project == null)
+            {
+                return NotFound();
+            }
+
+            if (project.UserId != GetUserId())
+            {
+                return Forbid();
+            }
+
             var projectDeleted = await _projectService.DeleteProjectAsync(id);
 
             if (!projectDeleted)
@@ -76,6 +104,11 @@ namespace ProjectManagementApp.Controllers
             }
 
             return NoContent();
+        }
+
+        private string GetUserId()
+        {
+            return _httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
         }
     }
 }
